Force configure-event synthesizing on start up (patch by Vladimir Murzin)
Transparency is handled via "configure-event" callback. However, if x
and y positions aren't changed (we get this case, for instance, with
GDK_GRAVITY_NORTH_WEST and edge top/left) event is not synthesized on
start up.

It seems that behavior of gtk_window_move() was updated with commit
6716d46393039c0a0691306d65a1e1921a6957fe (Bail out of window
move/resize early if no change) in gtk+.

Fix for that isn't elegant at all. We force initial position of panel
to GTK_WIN_POS_CENTER, so with any edge options panel has to be
moved and event will be generated.